requirements.txt
google-cloud-bigquery==1.28.0
google-cloud-storage==1.31.2
google-cloud-bigquery-storage==1.1.0
pyaml==20.4.0
Jinja2==2.11.2
tomlkit==0.7.0